make code structure easier to reason about
