+++ xain/generator/partition_volume_distributions.py
@@ -1,6 +1,5 @@
 from typing import Dict, List, Tuple
 
-import matplotlib.pyplot as plt
 import numpy as np
 
 bs_fashion_mnist: Dict[float, float] = {
@@ -58,18 +57,6 @@ def _generate_volume_distribution(xs: np.ndarray, a: float, b: float) -> List[in
     return [int(exponential_decay(x, a=a, b=b)) for x in xs]
 
 
-def plot_fashion_mnist_dist():
-    dists = fashion_mnist_100p()
-    xs = np.arange(100)
-    plt.figure()
-    legend = []
-    for b, dist in dists:
-        legend.append(str(b))
-        plt.plot(xs, np.array(dist))
-    plt.legend(legend, loc="upper left")
-    plt.show()
-
-
 def brute_force_a_for_fashion_mnist():
     for b in [1.0, 1.005, 1.01, 1.015, 1.02, 1.025, 1.03, 1.035, 1.04, 1.045]:
         a = brute_force_a(np.arange(100), b, target=54_000)
@@ -97,10 +84,20 @@ def brute_force_a(xs, b: float, target: int, step=1.0, start=1):
             return brute_force_a(xs, b, target, step / 10, start=a_best)
 
 
-if __name__ == "__main__":
-    print("Fashion-MNIST:")
-    brute_force_a_for_fashion_mnist()
-    print("CIFAR-10:")
-    brute_force_a_for_cifar_10()
-    print("Plot Fashion-MNIST volume distributions")
-    plot_fashion_mnist_dist()
+def b_to_str(b: float):
+    b_str = f"{b:<f}"
+    return b_str[:5].replace(".", "_")
+
+
+def dist_to_indicies(dist: List[int]) -> List[int]:
+    indices = [0] * len(dist)
+    for i, _ in enumerate(dist):
+        if i == 0:
+            indices[i] = dist[i]
+        else:
+            indices[i] = indices[i - 1] + dist[i]
+
+    assert indices[-1] == sum(dist)
+
+    # Exclude last element as indices only mark start of section
+    return indices[:-1]
