+++ xain/generator/partition_volume_distributions_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from . import volume_distributions
+from . import partition_volume_distributions as pvd
 
 
 def test_fashion_mnist_distributions():
@@ -9,7 +9,7 @@ def test_fashion_mnist_distributions():
     expected_sum = 54_000
 
     # Execute
-    dists = volume_distributions.fashion_mnist_100p()
+    dists = pvd.fashion_mnist_100p()
 
     # Assert
     assert len(dists) == 10
@@ -22,7 +22,7 @@ def test_cifar_10_distributions():
     expected_sum = 45_000
 
     # Execute
-    dists = volume_distributions.cifar_10_100p()
+    dists = pvd.cifar_10_100p()
 
     # Assert
     assert len(dists) == 10
@@ -43,6 +43,17 @@ def test_cifar_10_distributions():
 )
 def test_brute_force_a(b, expected, target):
     # Execute
-    actual = volume_distributions.brute_force_a(np.arange(100), b=b, target=target)
+    actual = pvd.brute_force_a(np.arange(100), b=b, target=target)
     # Assert
     assert actual == expected
+
+
+def test_dist_to_indicies():
+    # Prepare
+    dists = pvd.cifar_10_100p() + pvd.fashion_mnist_100p()
+
+    assert len(dists) == 20
+
+    for _, dist in dists:
+        indices = pvd.dist_to_indicies(dist)
+        assert len(indices) == len(dist) - 1
